fix: image error on build
The public images cannot be shown in the build version. After a thorough check of the codebase, we realize that the issue is because react does not use the /public folder in the build version. Therefore, we refactored the backend and use express to run a static resources provider for the frontend.